+++ src/components/CoinItemList/CoinItemList.tsx
@@ -1,18 +1,49 @@
+import React, { useEffect, useState, useCallback } from "react";
+
 import CoinItemContainer from "@components/CoinItemContainer/CoinItemContainer";
 import styles from "@components/CoinItemList/CoinItemList.module.scss";
+import axios from "axios";
 import { CoinsData } from "src/App";
+import { RawData } from "src/App";
 
 type CoinItemListData = {
+  setCoins: (coin: CoinsData[]) => void;
   coins: CoinsData[];
   category: string;
   isInputSearchOpen: boolean;
+  currency: string;
 };
 
 const CoinItemList = ({
+  setCoins,
+  currency,
   coins,
   category,
   isInputSearchOpen,
 }: CoinItemListData) => {
+  const fetch = useCallback(() => {
+    const fetch = async () => {
+      const result = await axios({
+        method: "get",
+        url: `https://api.coingecko.com/api/v3/coins/markets?vs_currency=${currency}`,
+      });
+      setCoins(
+        result.data.map((raw: RawData) => ({
+          id: raw.id,
+          image: raw.image,
+          name: raw.name,
+          symbol: raw.symbol,
+          curPrice: raw.current_price,
+          priceChange: raw.price_change_percentage_24h,
+          priceChangeFlat: raw.price_change_24h,
+        }))
+      );
+    };
+    fetch();
+  }, [currency]);
+  useEffect(() => {
+    fetch();
+  }, [currency]);
   let coinListClassNames = "coin__list__container_input_close";
   if (isInputSearchOpen === true) {
     coinListClassNames = "coin__list__container_input_open";
